Collapse all and multiple_*
